+++ extensions/mailwip/CHANGELOG.md
@@ -1,5 +1,8 @@
 # Mailwip Changelog
 
-## [] - {PR_MERGE_DATE}
+## [Better Error Handling] - {PR_MERGE_DATE}
+
+- update api to have more robust error handling
+- `optimistic update` for aliases through hook
 
 ## [Initial Version] - 2023-07-09